refactor(main): 一時ファイル存在確認を tempFile サービスに抽出
EXISTS_TEMP_FILE ハンドラのロジックを関数化し、IPC と main 内部
(modList サービス)の両方から使えるようにする。挙動は同一。
